# merged_canvas/canvas/openapi.yml
...
openapi: 3.1.0
info:
  title: Drupal Canvas
  description: API Spec for Drupal Canvas
  version: 0.x

# - Drupal's session auth (cookies) for all routes.
# - OAuth2 for some routes when the `canvas_oauth` submodule is installed.
# @see \Drupal\canvas_oauth\Routing\CanvasOauthRouteSubscriber:ROUTE_OPTION_EXTERNAL_API
security: []
...
paths:
...
  '/canvas/api/v0/content/{entity_type}':
    get:
      summary: Provides api for page listing.
      parameters:
        - in: query
          name: search
          required: false
          description: Search term to filter content by title
          schema:
            type: string
          examples:
            Find content with title 'My page':
              value: 'My page'
            Search for marketing update content:
              value: 'Marketing update'
      responses:
        200:
          description: Page list generated successfully.
          content:
            application/json:
              schema:
                type: object
                minProperties: 0
                # @todo Validate the property keys match the '^[0-9]$' pattern
                #   in https://drupal.org/i/3506543.
                additionalProperties:
                  type: object
                  required:
                    - id
                    - title
                    - status
                    - path
                    - autoSaveLabel
                    - autoSavePath
                    - internalPath
                    - links
                  properties:
                    id:
                      type: integer
                      description: The page ID.
                    title:
                      type: string
                      description: The page title.
                    status:
                      type: boolean
                      description: The page status.
                    autoSaveLabel:
                      type: [string, "null"]
                      description: The auto-save title.
                    autoSavePath:
                      type: [string, "null"]
                      format: uri-reference
                      description: The auto-save path.
                    internalPath:
                      type: string
                      format: uri-reference
                      description: The internal (unprocessed) path
                    path:
                      type: string
                      format: uri-reference
                      title: The processed path
                    links:
                      type: object
                      description: The operation links.
                      additionalProperties:
                        type: string
                      examples:
                        edit-form: /canvas/editor/canvas_page/3
              examples:
                pageList:
                  value:
                    '1':
                      id: 1
                      title: Home
                      status: true
                      path: /home
                      autoSaveLabel: Homer
                      autoSavePath: /updated-path
                      internalPath: /page/1
                      links:
                        edit-form: /canvas/editor/canvas_page/1
                    '2':
                      id: 2
                      title: About
                      status: false
                      path: /about
                      autoSaveLabel: About who?
                      autoSavePath: /page/1
                      internalPath: /page/2
                      links:
                        edit-form: /canvas/editor/canvas_page/2
    post:
      summary: Creates content for a specific entity type
      parameters:
        - name: entity_type
          in: path
          required: true
          description: The entity type
          schema:
            type: string
      requestBody:
        description: Optional entity ID for duplication
        content:
          application/json:
            schema:
              type: object
              properties:
                entity_id:
                  type: string
                  description: The ID of the entity to duplicate (optional)
      responses:
        '201':
          description: Created entity successfully
          content:
            application/json:
              examples:
                newContentEntity:
                  value:
                    entity_type: canvas_page
                    entity_id: '123'
              schema:
                type: object
                required:
                  - entity_type
                  - entity_id
                additionalProperties: false
                properties:
                  entity_type:
                    type: string
                    description: The entity type ID
                    # @todo https://www.drupal.org/i/3498525 should generalize this to all eligible content entity types
                    enum:
                      - canvas_page
                    examples:
                      - canvas_page
                  entity_id:
                    # @todo This is a bug https://www.drupal.org/i/3510925 introduced, fix it in https://www.drupal.org/i/3498525
                    type: string
                    description: The id of created content entity
                    examples:
                      - '42'
        '404':
          description: Entity not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Cannot find entity to duplicate.
...